hooked up calculator with web app
