+++ src/main/java/com/webCalc/springboot/demo/Stack.java
@@ -1,3 +1,6 @@
+package com.webCalc.springboot.demo;
+
+
 public class Stack<T> {
     public StackVar<T> topOfStack;
