+++ src/main/java/com/webCalc/springboot/demo/StackVar.java
@@ -1,3 +1,5 @@
+package com.webCalc.springboot.demo;
+
 public class StackVar<T> {
     private T item;
     public StackVar<T> nextVar;
